'feat: Refactor pull-request.yml and package.json
Refactored pull-request.yml file to run a "Check Pull Request" job that executes tests for Node versions 14, 16, and 18. The Pnpm package manager is being used for the test with the version defined by the PNPM_VERSION environment variable.
Additionally, package.json has been updated to include the build script for creating production files and adding the homepage and repository keys to the package information.'